Refactor HasQuestionnaireAnswers helpers
