+++ decidim-forms/app/helpers/decidim/forms/admin/concerns/has_questionnaire_answers_url_helper.rb
@@ -4,7 +4,15 @@ module Decidim
   module Forms
     module Admin
       module Concerns
-        module HasQuestionnaireAnswersHelpers
+        # Url helper for HasQuestionnaireAnswers controller concern
+        #
+        module HasQuestionnaireAnswersUrlHelper
+          def self.included(base)
+            base.helper_method :questionnaire_url, :questionnaire_participants_url,
+                               :questionnaire_participant_answers_url, :questionnaire_export_url,
+                               :questionnaire_export_response_url
+          end
+
           # You can implement this method in your controller to change the URL
           # where the questionnaire can be edited.
           def questionnaire_url
@@ -29,38 +37,6 @@ module Decidim
 
           def questionnaire_export_response_url(session_token)
             url_for([:export_response, questionnaire.questionnaire_for, session_token: session_token, format: "pdf"])
-          end
-
-          def prev_url
-            return if first?
-
-            token = participant_ids[current_idx - 1]
-            questionnaire_participant_answers_url(token)
-          end
-
-          def next_url
-            return if last?
-
-            token = participant_ids[current_idx + 1]
-            questionnaire_participant_answers_url(token)
-          end
-
-          def first?
-            current_idx.zero?
-          end
-
-          def last?
-            current_idx == participant_ids.count - 1
-          end
-
-          private
-
-          def participant_ids
-            @participant_ids ||= collection.pluck(:session_token)
-          end
-
-          def current_idx
-            participant_ids.index(params[:session_token])
           end
         end
       end
